Implement changes in the SIP signaling to enable encrypted conferences: - a RecvOnly audio stream with content DTX and the participant label for each participant of the conference. A secured conference will have N audio streams (1 sendrecv for the local participant and N-1 recvonly for the other participants) - add security level to the conference parameters - pass security level information to the conference server - add encryption information to the database table storing conferences - pass security level down to MSAudioConference and MSVideoConference objects - add extra video stream for encrypted conference in active speaker layout

+++ src/call/call.cpp
@@ -122,8 +122,8 @@ int Call::getMediaStreamIndex(LinphoneStreamType type) const {
 	return (int)s->getIndex();
 }
 
-int Call::getMediaStreamsNb(LinphoneStreamType type) const {
-	int nb = 0;
+size_t Call::getMediaStreamsNb(LinphoneStreamType type) const {
+	size_t nb = 0;
 	auto ms = static_pointer_cast<MediaSession>(getActiveSession())->getPrivate();
 	StreamsGroup &sg = ms->getStreamsGroup();
 	SalStreamType nType;
